add additional command buffer error checking using test queues (#127)
* add an option for command buffer emulation enhanced error checking

Adds an option for enhanced error checking, disabled by default.
Enhanced error checking creates special test queues when a command
buffer is created, and enqueues a barrier blocked by a user event
into the test queue.  Then, before a command is recorded into a
command buffer, it is also enqueued into the test queue, to identify
command errors.  When the command buffer is finalized the user event
is set to an error state, causing all of the commands in the test
queue to be terminated.

* fix typo

* add an environment variable to control enhanced error checking

* add checks for a few more command buffer errors